// src/types/wishList.type.ts
export interface WishListData {
    status: string;
    count:  number;
    data:   wishData[];
}

export interface wishData {
    sold:                number;
    images:              string[];
    subcategory:         Brand[];
    ratingsQuantity:     number;
    _id:                 string;
    title:               string;
    slug:                string;
    description:         string;
    quantity:            number;
    price:               number;
    priceAfterDiscount?: number;
    imageCover:          string;
    category:            Brand;
    brand:               Brand;
    ratingsAverage:      number;
    createdAt:           Date;
    updatedAt:           Date;
    __v:                 number;
    id:                  string;
    availableColors?:    unknown[];
}

export interface Brand {
    _id:       string;
    name:      string;
    slug:      string;
    image?:    string;
    category?: string;
}